Use method of Marrel et al. for stochastic metamodels
- fit model to mean, then dispersion, then mean with heteroscedasticiy, then dispersion
- use final two metamodels to compute sobol indices
- slight deviation: we use RMSE from a training set to determine homo. dispersion for
  initial model of mean

The documentation of DiceKriging seems to indicate that Marrel et al. used term nugget
for concept of observation noise, and that these are not the same thing. We use
observation noise and not nugget. Our deviation is due to the fact that DiceKriging
does not allow observation noise parameter to be estimated.

- Introduce some of the plotting methods from DiceView to view fits.