Add starting grid variables to design
Had to reduce multistart argument to 1 for some of the DiceEval calls in layer 4.
With 5 multistarts, an error occured in one of the optimisations and caused execution to halt. Not
clear if the error was due to lack of memory or something else.
The DiceKriging package could be patched here to tolerate an error in one of the optimisations:
https://github.com/cran/DiceKriging/blob/cf5f3c02b171cf3253ff29c23834d6ecaa78136d/R/kmEstimate.R#L137

fixes #13