Correct calculation of kriging residuals
Using prediction error was not valid. The residuals are now calculated
by using the predict method for a kriging model with observation noise
derived from an estimated homoscedastic nugget parameter or a Gaussian
process model for the squared residuals from a previous round of
fitting. Interesting to note, the squared residuals are much close to
Gaussian after this correction and thus modeling them with a Gaussian
process is much easier to justify. Also, the estimated dispersion
parameter seems a little low when fitting the joint metamodel to a few
hundred simulations. Perhaps with more data, and design points closer
together, that will change.