Stop creating cluster for single core processes
Creating a cluster will create another process. If available, the kriging
model fitting code will do jobs in this second process and at least from
the output of top it seems that memory usage is doubled. Since we'd like
to layer 4 in parallel for different response variables, we want to get
memory usage down to well under a GB. Thus this change.