Eliminate litter rate decrease from cum. burdens analysis
The latest version of glmnet is giving warnings that one of the
categories in the response has fewer than 8 observations and that we
are on dangerous ground. Also, stabbath() was running into errors
because some subsamples had fewer than 1 observation. Also, I'm
concerned that the assumption made that the litter rates of states
without individual litter rate estimates is close to the group average
may be contributing to the selection of the inventory variable. States
without individual rate estimates tend to be smaller, and the group
average for these states was small. Altogether, this assumption
probably has too much influence given that all of these states are
treated as independent observations. Possibly, something clever could
be done to address this problem, but as the result is of little
importance it seems more reasonable save my time and drop it.