Switch build system to use gcd response
Layers 4 and 5 now look for an environmental variable to determine
which response variable to choose. This makes it easy to launch a
different r job for each process. It By branching at the level of the
Makefile, we prevent the need for any single R process to keep the
results of all the different responses in memory. This way, the build
can occur on machines with only 1 GB of memory, and memory
requirements with the number of jobs active at any time rather than
the total number of jobs.