Remove kriging method from modelComparison methods
This should speed up the comparison considerably, since fitting the
kriging model to all 10 cross-validated datasets took some time. It
was mostly a waste too since the number of iterations for the bgfs
optimization could not be controlled and the iterations stopped before
convergence. We still fit the kriging model to the training set and
get prediction error from the test set separately and are able to see
how it does there compared to the other models.

This commit also deletes some comments in the GetMetaModels function
that were superfluous and liable to become inaccurate.

addresses: #17