Change design of SA
Fix target mean degree so as to shrink dimensionality a bit. It seems
as if fitting a Gaussian process to much more than 1000 points could
be burdensome, like an hour or so with bfgs. So it's not clear that we
can afford another dimension and the effect of the mean degree seems
fairly clear without including it. We still keep the raster cell area
varied because it would be interesting to show a point where it is
responsive to tprob.sp. At the 100 mi^2 setting, it looked like the
correlation saturated very quickly like before tprob.sp == 0.05. At
the 4 mi^2 setting, which might be more realistic for PEDV, there is
little sensitivity to tprob.sp.